Break out network tests, and replace sensitive data
